Create a calculatos file component and app component and style for all calculator
